feat: improve caching
