+++ src/events/messageCreate.ts
@@ -3,11 +3,11 @@ import { sendMessage } from '../commands/interaction/say';
 // import { sendVisitMessage } from '../commands/interaction/requestvisit';
 import serverModel from '../models/serverModel';
 import userModel from '../models/userModel';
-import { CustomClient, Event, ProxyConfigType, ProxyListType, Quid, ServerSchema, UserSchema } from '../typedef';
+import { CustomClient, DiscordEvent, ProxyConfigType, ProxyListType, Quid, ServerSchema, UserSchema } from '../typedef';
 import { getMapData } from '../utils/helperFunctions';
 import { createGuild } from '../utils/updateGuild';
 
-export const event: Event = {
+export const event: DiscordEvent = {
 	name: 'messageCreate',
 	once: false,
 	async execute(client: CustomClient, message: Message) {
