+++ src/paw.ts
@@ -1,21 +1,50 @@
-import { GatewayIntentBits, Partials } from 'discord.js';
-import { existsSync, writeFileSync } from 'fs';
-import { BanList, CustomClient, DeleteList, GivenIdList, VoteList, WebhookMessages } from './typedef';
+import { Collection, GatewayIntentBits, LimitedCollection } from 'discord.js';
+import { existsSync, readdirSync, readFileSync, writeFileSync } from 'fs';
+import { BanList, CustomClient, DeleteList, DiscordEvent, GivenIdList, UserSchema, VoteList, WebhookMessages } from './typedef';
+import path from 'path';
 
 /* Note: Once slash commands replace message commands, DIRECT_MESSAGES intent and CHANNEL partial can be removed */
 export const client = new CustomClient({
 	intents: [
 		GatewayIntentBits.Guilds,
 		GatewayIntentBits.GuildMessages,
-		GatewayIntentBits.DirectMessages,
 		GatewayIntentBits.MessageContent,
-	],
-	partials: [
-		Partials.Channel,
 	],
 	allowedMentions: {
 		parse: ['users', 'roles'],
 		repliedUser: false,
+	},
+	makeCache: manager => {
+		if (manager.name === 'GuildMemberManager') {
+			return new LimitedCollection({ maxSize: 0, keepOverLimit: (member) => {
+				const allDocumentNames = readdirSync('./database/profiles').filter(f => f.endsWith('.json'));
+				return allDocumentNames
+					.map(documentName => {
+						return JSON.parse(readFileSync(`./database/profiles/${documentName}`, 'utf-8')) as UserSchema;
+					})
+					.filter(v => v.userId.includes(member.id))
+					.length > 0;
+			} });
+		}
+		if (
+			manager.name === 'ApplicationCommandManager'
+			|| manager.name === 'BaseGuildEmojiManager'
+			|| manager.name === 'GuildBanManager'
+			|| manager.name === 'GuildEmojiManager'
+			|| manager.name === 'GuildInviteManager'
+			|| manager.name === 'GuildScheduledEventManager'
+			|| manager.name === 'GuildStickerManager'
+			|| manager.name === 'MessageManager'
+			|| manager.name === 'PresenceManager'
+			|| manager.name === 'ReactionManager'
+			|| manager.name === 'ReactionUserManager'
+			|| manager.name === 'StageInstanceManager'
+			|| manager.name === 'ThreadManager'
+			|| manager.name === 'ThreadMemberManager'
+			|| manager.name === 'UserManager'
+			|| manager.name === 'VoiceStateManager'
+		) { return new LimitedCollection({ maxSize: 0 }); }
+		return new Collection();
 	},
 });
 
@@ -65,11 +94,25 @@ export async function start(
 		writeFileSync('./database/webhookCache.json', JSON.stringify(({}) as WebhookMessages, null, '\t'));
 	}
 
-	/* It's loading all the files in the handlers folder. */
-	for (const file of ['commands', 'events', 'votes']) {
+	for (const file of readdirSync(path.join(__dirname, './events'))) {
 
-		try { await require(`./handlers/${file}`).execute(client); }
-		catch (error) { console.error(error); }
+		console.log(file);
+		const { event } = require(`./events/${file}`) as { event: DiscordEvent; };
+
+		if (event.once) {
+
+			client.once(event.name, (...args) => {
+				try { event.execute(client, ...args); }
+				catch (error) { console.error(error); }
+			});
+		}
+		else {
+
+			client.on(event.name, (...args) => {
+				try { event.execute(client, ...args); }
+				catch (error) { console.error(error); }
+			});
+		}
 	}
 
 	await client.login(botToken);
